requirements.txt
# Core Dependencies
anthropic>=0.40.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Gmail API
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# Notion API
notion-client>=2.2.0

# Storage & Analysis
sqlalchemy>=2.0.0
pandas>=2.0.0

# CLI
click>=8.1.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
